+++ view/src/components/Wizard/allWizards.tsx
@@ -3,11 +3,13 @@ import DeleteElementsWizard from './DeleteElementsWizard/DeleteElementsWizard';
 import CreateFaultContainmentRegionWizard from './CreateFCRWizard';
 import LifecycleStateWizard from './LifecycleStateWizard';
 import UpdateElementsWizard from './UpdateElementsWizard/UpdateElementsWizard';
+import RelationElementsWizard from './RelationElementsWizard/RelationElementsWizard';
 
 export default {
   DeleteElementsWizard,
   CreateFaultContainmentRegionWizard,
   LifecycleStateWizard,
   CreateElementWizard,
-  UpdateElementsWizard
+  UpdateElementsWizard,
+  RelationElementsWizard
 };